Fix more windows build warnings.
TBR=ulan@chromium.org
BUG=

Review URL: https://chromiumcodereview.appspot.com/10917065

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@12433 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/api.cc
@@ -4961,7 +4961,8 @@ Local<String> v8::String::NewExternal(ExternalLatin1StringResource* resource,
     // since it's not.  Instead, we convert it to an internal string and dispose
     // the external resource.
     result = isolate->factory()->NewStringFromLatin1(
-        i::Vector<const char>(resource->data(), resource->length()),
+        i::Vector<const char>(resource->data(),
+                              static_cast<int>(resource->length())),
         i::NOT_TENURED,
         i::String::NOT_ASCII);
     resource->Dispose();
